Optimize TruthTable.split and .merge to do nothing in common case
Most TruthTables have a default of all dont care. We can optimize this
common case by doing actual work in TruthTable.split and
TruthTable.merge.
